% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{filter_sfa}
\alias{filter_sfa}
\title{Filter variable}
\usage{
filter_sfa(df, y, x_list = NULL, missing_limit = 0.95,
  iv_limit = 0.01, auc_limit = 0.5, ks_limit = 0.1,
  identical_limit = 0.95, var_skip = NULL, return_rm_reason = TRUE)

filter_sfa(df, y, x_list = NULL, missing_limit = 0.95,
  iv_limit = 0.01, auc_limit = 0.5, ks_limit = 0.1,
  identical_limit = 0.95, var_skip = NULL, return_rm_reason = TRUE)
}
\arguments{
\item{df}{data.frame contains y and x_list variables}

\item{y}{binary variable}

\item{x_list}{list predictor variables}

\item{missing_limit}{The missing rate of kept variables should <= missing_limit. The default is 0.95.}

\item{iv_limit}{The information value of kept variables should >= iv_limit. The default is 0.02.}

\item{auc_limit}{The AUC of kept variables should >= auc_limit The default is 0.5.}

\item{ks_limit}{The Kolmogorov Smirnov of kept variables should >= ks_limit The default is 0.1.}

\item{identical_limit}{The identical value rate (excluding NAs) of kept variables should <= identical_limit. The default is 0.95.}

\item{var_skip}{Name of force kept variables, default is NULL.}

\item{return_rm_reason}{Logical, default is FALSE.}

\item{y}{binary variable}

\item{x_list}{list predictor variables}
}
\value{
data.frame include pass variables

data.frame include pass variables
}
\description{
This function filter variables base on specified conditions, such as information value, missing rate, identical value rate, Kolmogorov Smirnov statistic.

This function filter variables base on specified conditions, such as information value, missing rate, identical value rate, Kolmogorov Smirnov statistic.
}
\examples{
data('hmeq')
hmeq1 <- hmeq[,c("bad", "reason", "job")]
x <- filter_sfa(hmeq1,
           y = 'bad',
           missing_limit = 0.95,
           iv_limit = 0.3,
           auc_limit = 0.5,
           ks_limit = 0.4,
           identical_limit = 0.95,
           var_skip = 'loan',
           return_rm_reason = TRUE )
data('hmeq')
x <- var_filter(hmeq,
           y = "bad",
           missing_limit = 0.95,
           iv_limit = 0.3,
           auc_limit = 0.5,
           ks_limit = 0.4,
           identical_limit = 0.95,
           var_rm = NULL,
           var_skip = "loan",
           return_rm_reason = TRUE )
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}

Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
