% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{get_auc}
\alias{get_auc}
\title{get_auc}
\usage{
get_auc(x, y)
}
\arguments{
\item{x}{categorical variable}

\item{y}{binary variable}
}
\value{
AUC value
}
\description{
Compute the auc
}
\details{
With a Binary Y variable and a categorical X variable stored as factor. AUC value will be calculated
}
\examples{
data('hmeq')
auc1 <- get_auc(x = hmeq$reason, y = hmeq$bad)
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
