% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{get_ks}
\alias{get_ks}
\title{Compute Kolmogorov Smirnov statistic}
\usage{
get_ks(x, y)
}
\arguments{
\item{x}{categorical variable}

\item{y}{binary variable}
}
\value{
AUC value
}
\description{
Compute the ks
}
\details{
With a Binary Y variable and a categorical X variable stored as factor. KS value will be calculated
}
\examples{
data('hmeq')
ks1 <- get_ks(x = hmeq$reason, y = hmeq$bad)
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
