% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfa.R
\name{get_missing}
\alias{get_missing}
\title{Compute missing rate}
\usage{
get_missing(x)

get_missing(x)
}
\arguments{
\item{x}{input variable}
}
\description{
Compute missing rate

Compute missing rate
}
\examples{
data('hmeq')
get_missing(hmeq$loan)
data('hmeq')
get_missing(hmeq$loan)
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}

Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
