% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{psi}
\usage{
psi(original, current, cut.points = NULL,
  cut.levels = ifelse(is.null(cut.points), 5L, NULL))

psi(original, current, cut.points = NULL,
  cut.levels = ifelse(is.null(cut.points), 5L, NULL))
}
\arguments{
\item{original}{The original set of a measurement, should be a factor or numeric}

\item{current}{The current set of a measurement, should be a factor or numeric}

\item{cut.points}{It won't work if original and current are factors, and it cannot be NULL if original and current are numerical. This function uses this argument to bin \code{original} and \code{current} with left-closed right-open intervals.}

\item{cut.levels}{specifies how many levels to split a numeric variable. If \code{cut.points} is provided, this parameter will be ignored. When using this parameter, a numeric variable will be split using \code{cut}.}

\item{x_train}{expected or training variable, should be a factor}

\item{x_test}{observed or actual variable, should be a factor}
}
\value{
psi value

a \code{psi} object
}
\description{
Compute the psi

\code{psi} calculates the popolation stability index.
}
\details{
Population stability index (PSI) or system stability index (SSI)

Check whether the population on which the model is currently being used is similar to the population that was used to develop it.

psi measures the stablity of the population. Usually we can believe the population stays the same as the past if psi is less than 0.1, and a significant shift can be recognised if psi is greater than 0.25. The outcome of this function is a numeric, with details stored as attributes. You can use \code{summary} function to see all of the detailed information. Fot the situation where some of the levels has no element in either original population or current population and the psi does not exist for such levels, the empty levels will not be taken into account and a warning will inform you of this. Again, by using \code{summary} you could know everything inside.
}
\examples{
data('hmeq')
library(rsample)
hmeq1 <- hmeq[,c("bad", "reason", "job")]
spl <- initial_split(hmeq1)
train_df <- training(spl)
test_df <- testing(spl)
psi1 <- psi(train_df$job, test_df$job)

data('iris')
train <- sample(nrow(iris), nrow(iris) * .7)
train.species <- iris$Species[train]
test.species <- iris$Species[-train]
p <- psi(train.species, test.species)
p
summary(p)

}
\references{
P395, Credit risk analytics - Measurement technique, applications and examples
Rules of thumb when using psi:
\tabular{rr}{
PSI \tab shift \cr
----------------- \tab ---------------- \cr
     < 0.1 \tab no significant shift (green traffic light) \cr
0.1 to 0.25 \tab moderate shift (yellow traffic light) \cr
      > 0.25 \tab significant shift (red traffic light)
}

\href{What is population stability index? - Quora}{https://www.quora.com/What-is-population-stability-index}
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}

Siyuan Yao
}
