% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi_df}
\alias{psi_df}
\title{psi_df}
\usage{
psi_df(dat_train, dat_test, var_skip = NULL)
}
\arguments{
\item{var_skip}{Name of force kept variables, Defaults to NULL}

\item{df_train}{expected or training variable, should be a factor}

\item{df_test}{observed or actual variable, should be a factor}
}
\value{
psi value
}
\description{
Compute the psi for all dataset
}
\details{
Check whether the population on which the model is currently being used is similar to the population that was used to develop it.
}
\examples{
data('hmeq')
library(rsample)
hmeq1 <- hmeq[,c("bad", "reason", "job")]
spl <- rsample::initial_split(hmeq1)
train_df <- training(spl)
test_df <- testing(spl)
psi_df1 <- psi_df(train_df, test_df)
}
\references{
P395, Credit risk analytics - Measurement technique, applications and examples
Rules of thumb when using psi:
\tabular{rr}{
PSI \tab shift \cr
----------------- \tab ---------------- \cr
     < 0.1 \tab no significant shift (green traffic light) \cr
0.1 to 0.25 \tab moderate shift (yellow traffic light) \cr
      > 0.25 \tab significant shift (red traffic light)
}
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
