% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_function.R
\name{tab}
\alias{tab}
\title{tab}
\usage{
tab(vec)
}
\arguments{
\item{vec}{is input vector}
}
\value{
frequency table
}
\description{
Quickly show frequency table
}
\details{
Quickly show frequency table
}
\examples{
tab(iris$Species)
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
