% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_function.R
\name{vietnamese_code}
\alias{vietnamese_code}
\title{vietnamese_code}
\usage{
vietnamese_code(show = TRUE)
}
\arguments{
\item{show}{is logic option}
}
\value{
vietnamese hex text
}
\description{
Show encode vietnamese
}
\details{
Show encode vietnamese
}
\examples{
vietnamese_code()
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
