% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe_cal_df.R
\name{woe_cal_quantiles}
\alias{woe_cal_quantiles}
\title{woe_cal_quantiles}
\usage{
woe_cal_quantiles(x, y, valueOfBad = 1, nbr_bin = 20)
}
\arguments{
\item{x}{The categorical variable stored as factor for which WOE Table is to be computed.}

\item{y}{The actual 1/0 flags for the binary response variable. It can take values of either 1 or 0, where 1 represents the 'Bad' or 'Events' while 0 represents 'Good' or 'Non-Events'.}

\item{valueOfBad}{The value in y that is used to represent 'Bad' or the occurence of the event of interest. Defaults to 1.}

\item{nbr_bin}{default = 20}
}
\value{
The WOE table with the respective weights of evidence for each group and the IV's.
\itemize{
  \item group The groups (levels) of the categorical X variable for which WOE is to be calculated.
  \item total. The total number of observations in respective group.
  \item good. The total number of 'Goods' or 'Events' in respective group.
  \item bad. The total number of 'Bads' or 'Non-Events' in respective group.
  \item pert. The Percentage of observations accounted for by respective group.
  \item pertg. The Percentage of 'Goods' or 'Events' accounted for by respective group.
  \item pertb. The Percentage of 'Bads' or 'Non-Events' accounted for by respective group.
  \item br. The Percentage of 'Bads' or 'Non-Events' in respective group.
  \item woe. The computed weights of evidence(WOE) for respective group. The WOE values can be used in place of the actual group itself, thereby producing a 'continuous' alternative.
  \item iv. The information value contributed by each group in the X. The sum of IVs is the total information value of the categorical X variable.
}
}
\description{
Compute the woe_cal_quantiles that shows the Weights Of Evidence (WOE) for each group and respective Information Values (IVs). The group X is devided by quantile
}
\details{
For a given actual for a Binary Y variable and a numeric X variable, the WOE table is generated with calculated WOE's and IV's
}
\examples{
data('hmeq')
woe_cal_quantiles(x = hmeq$loan, y = hmeq$bad)
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
