% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{psi}
\usage{
psi(x_train, x_test)
}
\arguments{
\item{x_train}{expected or training variable, should be a factor}

\item{x_test}{observed or actual variable, should be a factor}
}
\value{
psi value
}
\description{
Compute the psi
}
\details{
Population stability index (PSI) or system stability index (SSI)

Check whether the population on which the model is currently being used is similar to the population that was used to develop it.
}
\examples{
data('hmeq')
library(rsample)
hmeq1 <- hmeq[,c("bad", "reason", "job")]
spl <- initial_split(hmeq1)
train_df <- training(spl)
test_df <- testing(spl)
psi1 <- psi(train_df$job, test_df$job)

}
\references{
P395, Credit risk analytics - Measurement technique, applications and examples
Rules of thumb when using psi:
\tabular{rr}{
PSI \tab shift \cr
----------------- \tab ---------------- \cr
     < 0.1 \tab no significant shift (green traffic light) \cr
0.1 to 0.25 \tab moderate shift (yellow traffic light) \cr
      > 0.25 \tab significant shift (red traffic light)
}
}
\author{
Nguyen Ngoc Binh \email{nguyenngocbinhneu@gmail.com}
}
